Moved apiRoute file to api/index.ts
