+++ src/api/index.ts
@@ -1,1 +1,8 @@
-export { default as sample } from './sample'
+import express from 'express'
+import sample from './sample'
+
+const router = express.Router()
+
+router.get('/sample', sample)
+
+export default router
