--- src/apiRoute.ts
@@ -1,9 +1,0 @@
-import express from 'express'
-import { sample } from './api'
-
-const router = express.Router()
-
-router.get('/sample', sample)
-
-
-export default router
